+++ nn/layer.go
@@ -14,8 +14,8 @@ func NewLayer(numNeurons, numInputs int) *Layer {
 	}
 }
 
-func (l *Layer) Process(inputs []float64) []float64 {
-	outputs := make([]float64, len(l.Neurons))
+func (l *Layer) Process(inputs []float32) []float32 {
+	outputs := make([]float32, len(l.Neurons))
 	for i, neuron := range l.Neurons {
 		outputs[i] = neuron.Process(inputs)
 	}
